Split cli/lib/server-files.ts into appdata.ts + dependency-management/paths.ts
