+++ apps/cli/lib/appdata.ts
@@ -1,0 +1,20 @@
+import os from 'os';
+import path from 'path';
+import { __ } from '@wordpress/i18n';
+import { LoggerError } from 'cli/logger';
+
+export function getAppdataDirectory(): string {
+	if ( process.env.E2E && process.env.E2E_APP_DATA_PATH ) {
+		return path.join( process.env.E2E_APP_DATA_PATH, 'Studio' );
+	}
+
+	if ( process.platform === 'win32' ) {
+		if ( ! process.env.APPDATA ) {
+			throw new LoggerError( __( 'Studio config file path not found.' ) );
+		}
+
+		return path.join( process.env.APPDATA, 'Studio' );
+	}
+
+	return path.join( os.homedir(), 'Library', 'Application Support', 'Studio' );
+}
