+++ apps/cli/lib/site-language.ts
@@ -3,8 +3,8 @@ import path from 'path';
 import { Locale } from '@formatjs/intl-locale';
 import { match } from '@formatjs/intl-localematcher';
 import { DEFAULT_LOCALE } from '@studio/common/lib/locale';
+import { getWpFilesPath } from 'cli/lib/dependency-management/paths';
 import { getAppLocale } from 'cli/lib/i18n';
-import { getWpFilesPath } from 'cli/lib/server-files';
 
 interface TranslationsData {
 	translations: Translation[];
